third_party/blink/web_tests/editing/inserting/insert-paragraph-copy-style.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
function runTest(style_text) {
  test(() => {
    assert_selection(
      `<div contenteditable><div style="${style_text}">a|</div></div>`,
      'insertParagraph',
      `<div contenteditable><div style="${style_text}">a</div>` +
      `<div style="${style_text}">|<br></div></div>`);
  }, `${style_text}`);
}

const tests = [
  'text-decoration: underline;',
  'white-space: normal;',
  'white-space: pre;',
  'text-wrap: balance;',
];
for (const test of tests) {
  runTest(test);
}
</script>
